Integrate guide profile and tour management flows with the backend

+++ app/src/main/java/com/ahmetkaragunlu/guidemate/profile/domain/model/performance/GuidePerformanceSummary.kt
@@ -3,9 +3,9 @@ package com.ahmetkaragunlu.guidemate.profile.domain.model.performance
 import com.ahmetkaragunlu.guidemate.profile.domain.model.level.GuideLevelTier
 
 data class GuidePerformanceSummary(
-    val completedSessionCount: Int,
-    val totalParticipantCount: Int,
+    val completedSessionCount: Long,
+    val totalParticipantCount: Long,
     val averageRating: Double,
-    val reviewCount: Int,
+    val reviewCount: Long,
     val level: GuideLevelTier,
 )
